% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm - shortest node-to-node distances (base R)
Ref: https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{data frame with three variables:
v1, v2 = from and to nodes respectively (integers)
w = weight of edge}

\item{init_node}{(integer, present in graph)}
}
\value{
vector with the shortest distances from init_node to all nodes
}
\description{
Dijkstra algorithm - shortest node-to-node distances (base R)
Ref: https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm
}
