% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra extra.R
\name{dijkstra_dplyr_pipe}
\alias{dijkstra_dplyr_pipe}
\title{Dijkstra algorithm - shortest node-to-node distances (dplyr-pipe)
(dplyr pipe version)
Ref: https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm}
\usage{
dijkstra_dplyr_pipe(graph, init_node)
}
\arguments{
\item{graph}{data frame with three variables:
v1, v2 = from and to nodes respectively (integers)
w = weight of edge}

\item{init_node}{(integer, present in graph)}
}
\value{
vector with the shortest distances from init_node to all nodes
}
\description{
Dijkstra algorithm - shortest node-to-node distances (dplyr-pipe)
(dplyr pipe version)
Ref: https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
