% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra extra.R
\name{dijkstra_full}
\alias{dijkstra_full}
\title{Dijkstra algorithm finding the shortest node-to-node distances
(dplyr pipe version)
Ref: https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm}
\usage{
dijkstra_full(graph, init_node)
}
\arguments{
\item{graph}{data frame with three variables:
v1, v2 = from and to nodes respectively (integers)
w = weight of edge}

\item{init_node}{(integer, present in graph)}
}
\value{
data frame with full indo on nodes, i.e.,
node name, distance from init_node, and previous node
}
\description{
Dijkstra algorithm finding the shortest node-to-node distances
(dplyr pipe version)
Ref: https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
