% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest common divisor using the Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar representing an integer.}

\item{b}{A numeric scalar representing an integer.}
}
\value{
The greatest common divisor of `a` and `b`.
}
\description{
Computes the greatest common divisor of two integers using
the Euclidean algorithm.
}
\examples{
euclidean(100, 1000)
euclidean(123612, 13892347912)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
