% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph dataset}
\format{
A data frame with 3 columns:
\describe{
  \item{v1}{From node (integer).}
  \item{v2}{To node (integer).}
  \item{w}{Weight of the edge (numeric).}
}
}
\usage{
wiki_graph
}
\description{
This dataset is a sample graph used for testing the Dijkstra algorithm implementation. It contains nodes and edges with weights, representing a simple directed graph.
}
\keyword{datasets}
